wip toward zarr v2 reader
