+++ virtualizarr/readers/__init__.py
@@ -4,7 +4,7 @@ from virtualizarr.readers.hdf5 import HDF5VirtualBackend
 from virtualizarr.readers.kerchunk import KerchunkVirtualBackend
 from virtualizarr.readers.netcdf3 import NetCDF3VirtualBackend
 from virtualizarr.readers.tiff import TIFFVirtualBackend
-from virtualizarr.readers.zarr_v3 import ZarrV3VirtualBackend
+from virtualizarr.readers.zarrV2V3 import ZarrVirtualBackend
 
 __all__ = [
     "DMRPPVirtualBackend",
@@ -13,5 +13,5 @@ __all__ = [
     "KerchunkVirtualBackend",
     "NetCDF3VirtualBackend",
     "TIFFVirtualBackend",
-    "ZarrV3VirtualBackend",
+    "ZarrVirtualBackend",
 ]
